move semantic version into app class
